==> 02_word2vec_training.ipynb <==
import pandas as pd
import numpy as np
import csv
from gensim.test.utils import common_texts
from gensim.models import Word2Vec

with open('causenet_with_counts.csv', 'r',encoding='UTF-8') as data:
    reader = csv.DictReader(data)
    items = list(reader)
type(items)

sentence = []        
for x in items:
    cause = x.get('cause')
    effect = x.get('effect')
    obj = [cause,effect]
    sentence.append(obj)
    
model = Word2Vec(sentences=sentence, size=300, window=1, min_count=1, workers=4)    

model.save("scratch/word2vec.model")

model = Word2Vec.load("scratch/word2vec.model")

model.train([x for x in sentence], total_examples=1, epochs=1)

model.wv['accident']
